requirements.txt
Flask==3.0.0
PyMuPDF==1.24.0
Pillow==10.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
PyJWT==2.8.0
python-dotenv==1.0.0
